modified save_access_key()
