+++ src/commands/account/add_key/autogenerate_new_keypair/save_keypair_to_macos_keychain/mod.rs
@@ -14,9 +14,11 @@ impl SaveKeypairToMacosKeychain {
         prepopulated_unsigned_transaction: near_primitives::transaction::Transaction,
     ) -> crate::CliResult {
         let network_config = self.network_config.get_network_config(config.clone());
+        let key_pair_properties_buf = serde_json::to_string(&key_pair_properties)?;
         crate::common::save_access_key_to_macos_keychain(
             network_config,
-            key_pair_properties,
+            &key_pair_properties_buf,
+            &key_pair_properties.public_key_str,
             &prepopulated_unsigned_transaction.receiver_id,
         )
         .map_err(|err| {
