+++ src/commands/account/import_account/mod.rs
@@ -1,5 +1,8 @@
+use inquire::{CustomType, Select};
+use std::{str::FromStr, vec};
 use strum::{EnumDiscriminants, EnumIter, EnumMessage};
 
+// mod from_private_key;
 mod from_web_wallet;
 
 #[derive(Debug, Clone, interactive_clap::InteractiveClap)]
@@ -24,13 +27,122 @@ pub enum ImportAccountActions {
         message = "import-account-from-web-wallet          - Import existing account (a.k.a. \"sign in\")"
     ))]
     /// Import existing account (a.k.a. "sign in")
-    ImportAccountFromWebWallet(self::from_web_wallet::Login),
+    ImportAccountFromWebWallet(self::from_web_wallet::LoginFromWebWallet),
+    // #[strum_discriminants(strum(
+    //     message = "import-account-from-private-key         - Import existing account from private key"
+    // ))]
+    // /// Import existing account from private key
+    // ImportAccountFromPrivateKey(self::from_private_key::LoginFromPrivateKey),
 }
 
 impl ImportAccountActions {
     pub async fn process(&self, config: crate::config::Config) -> crate::CliResult {
         match self {
             Self::ImportAccountFromWebWallet(login) => login.process(config).await,
+            // Self::ImportAccountFromPrivateKey(login) => login.process(config).await,
         }
     }
+}
+
+pub async fn login(
+    network_config: crate::config::NetworkConfig,
+    credentials_home_dir: std::path::PathBuf,
+    key_pair_properties: crate::common::KeyPairProperties,
+    error_message: String,
+) -> crate::CliResult {
+    let public_key: near_crypto::PublicKey =
+        near_crypto::PublicKey::from_str(&key_pair_properties.public_key_str)?;
+
+    let account_id = loop {
+        let account_id_from_cli = input_account_id()?;
+        println!();
+        if crate::common::verify_account_access_key(
+            account_id_from_cli.clone(),
+            public_key.clone(),
+            network_config.clone(),
+        )
+        .await
+        .is_err()
+        {
+            println!("{}", error_message);
+            #[derive(strum_macros::Display)]
+            enum ConfirmOptions {
+                #[strum(to_string = "Yes, I want to re-enter the account_id.")]
+                Yes,
+                #[strum(to_string = "No, I want to save the access key information.")]
+                No,
+            }
+            let select_choose_input = Select::new(
+                "Would you like to re-enter the account_id?",
+                vec![ConfirmOptions::Yes, ConfirmOptions::No],
+            )
+            .prompt()?;
+            if let ConfirmOptions::No = select_choose_input {
+                break account_id_from_cli;
+            }
+        } else {
+            break account_id_from_cli;
+        }
+    };
+    save_access_key(
+        account_id,
+        key_pair_properties,
+        network_config,
+        credentials_home_dir,
+    )?;
+
+    Ok(())
+}
+
+fn input_account_id() -> color_eyre::eyre::Result<near_primitives::types::AccountId> {
+    Ok(CustomType::new("Enter account ID").prompt()?)
+}
+
+fn save_access_key(
+    account_id: near_primitives::types::AccountId,
+    key_pair_properties: crate::common::KeyPairProperties,
+    network_config: crate::config::NetworkConfig,
+    credentials_home_dir: std::path::PathBuf,
+) -> crate::CliResult {
+    let key_pair_properties_buf = serde_json::to_string(&key_pair_properties)?;
+
+    #[cfg(target_os = "macos")]
+    {
+        let macos_keychain = "Store the access key in my macOS keychain";
+        let legacy_keychain =
+            "Store the access key in my legacy keychain (compatible with the old near CLI)";
+        let selection = Select::new(
+            "Select a keychain to save the access key to:",
+            vec![macos_keychain, legacy_keychain],
+        )
+        .prompt()?;
+        if selection == macos_keychain {
+            let storage_message = crate::common::save_access_key_to_macos_keychain(
+                network_config,
+                &key_pair_properties_buf,
+                &key_pair_properties.public_key_str,
+                &account_id,
+            )
+            .map_err(|err| {
+                color_eyre::Report::msg(format!(
+                    "Failed to save the access key to the keychain: {}",
+                    err
+                ))
+            })?;
+            println!("{}", storage_message);
+            return Ok(());
+        }
+    }
+    let storage_message = crate::common::save_access_key_to_keychain(
+        network_config,
+        credentials_home_dir,
+        &key_pair_properties_buf,
+        &key_pair_properties.public_key_str,
+        &account_id,
+    )
+    .map_err(|err| {
+        color_eyre::Report::msg(format!("Failed to save a file with access key: {}", err))
+    })?;
+    println!("{}", storage_message);
+    Ok(())
 }
